+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/components/userBar.tsx
@@ -94,6 +94,7 @@ export default function UserBar(props: UserBarProps) {
             >
                 <img
                     src={DefaultUserIcon}
+                    alt="Open user menu"
                     height={28}
                     width={28}
                     onDragStart={event => {