+++ app/ide-desktop/lib/dashboard/src/authentication/src/dashboard/localBackend.ts
@@ -3,6 +3,8 @@
  * Each exported function in the {@link LocalBackend} in this module corresponds to an API endpoint.
  * The functions are asynchronous and return a {@link Promise} that resolves to the response from
  * the API. */
+import * as detect from 'enso-common/src/detect'
+
 import * as backend from './backend'
 import * as dateTime from './dateTime'
 import * as errorModule from '../error'
@@ -33,7 +35,7 @@ export class LocalBackend extends backend.Backend {
     constructor(projectManagerUrl: string | null) {
         super()
         this.projectManager = projectManager.ProjectManager.default(projectManagerUrl)
-        if (IS_DEV_MODE) {
+        if (detect.IS_DEV_MODE) {
             // @ts-expect-error This exists only for debugging purposes. It does not have types
             // because it MUST NOT be used in this codebase.
             window.localBackend = this